Delete README.md
There the file on which the raw VBA code was written and is working extremely well. When players in the prequalification tables have the same amount of wins and losses the algorithm first checks the points acquired trough sets and if the result is the same, it goes a spet further and calculated the points accumulated in each sets, and then changes the position accordingly to players place in the table. 
Comments are still in Slovene, since this file was originally created by the NTZS (Slovene table tennis association), I just added the behind the scene VBA code to it. in case you need to know of some tricks in VBA, there are plenty of challenges to work with. 

I hope you enjoy it and make your life with VBA a bit less stressful. A bigger project regarding the same idea is in the making where excellent players are used.